Add workflow steps to handle status and commits (#3)
The agent workflow in `.gitea/workflows/agent.yaml` writes the `.pr/history.md` file and runs the agent but it does not currently handle the agent's output correctly.

Specifically, the agent expects the following to happen after it completes running:
* If a `.pr/status.md` file is written, the contents will be posted as a comment on the PR
* If there are changes in the repository, they are committed with the commit message in `.pr/commit.md` and pushed to the commit branch automatically attached to the "laforge" user ID
* A `.pr/commit.md` file can be omitted if the only changes are in the `.pr` directory; we can generate a summary automatically
* Special files `.pr/history.md`, `.pr/status.md` and `.pr/commit.md` are never included in the commit

We also need to be careful that the commit pushed to the PR does not trigger the agent to run again in an infinite loop.

Co-authored-by: Tom W. Yedwab <tom@khanacademy.org>
Co-committed-by: Tom W. Yedwab <tom@khanacademy.org>