Use "Assignees" rather than "Reviewers" to loop in bot
Apparently once the laforge agent has left a "review" (which happens anytime it wants to leave file-level comments) it can no longer be removed as a "reviewer" and so will continue to respond to comments until the PR is closed. It looks like we should instead trigger the workflow in .gitea/workflows if laforge is set as an _assignee_, which hopefully can be added or removed at any time. This should also be controllable by the "unassign" field in status.yaml to allow the bot to unassign itself.

Co-authored-by: Tom W. Yedwab <tom@khanacademy.org>
Co-committed-by: Tom W. Yedwab <tom@khanacademy.org>