better links (#12)
Currently there are instructions in the claude instructions file for how to Output links to files in PR status messages however these link to the pr branch which can get out of date after the status messages written. instead we should link to a URL that includes the commit Sha but the complication is that the commit hasn't been created when the agent is generating the status message so the status message should include a placeholder which will be replaced in the update status script with the final commit Sha.

Co-authored-by: Laforge agent <laforge@tomyedwab.com>
Reviewed-on: http://localhost:3010/tom/laforge-2/pulls/12