enable agent selection (#16)
We may want to support multiple agents running in the pr workflow, either so I can use different models on a case by case basis to implement the pr, or so I can have multiple models critiquing the implementation.  I am not sure how to do this in the context of Gitea actions however.  Let's brainstorm some options.

Co-authored-by: Laforge agent <laforge@tomyedwab.com>
Reviewed-on: http://localhost:3010/tom/laforge-2/pulls/16
Co-authored-by: Tom W. Yedwab <tom@khanacademy.org>
Co-committed-by: Tom W. Yedwab <tom@khanacademy.org>